Fix int32 overflow in PBC CUDA kernel pointer arithmetic
For periodic cells with >~1000 basis functions, pointer offsets like
(ao_loc * nGv + Gv_id) * OF_COMPLEX exceed INT32_MAX, causing
CUDA_ERROR_ILLEGAL_ADDRESS. Cast to size_t before multiplication.

Affects ft_ao.cu, ft_ao_ip1.cu, nr_eval_gto.cu, mole_helper.cu.
Verified with compute-sanitizer on 128-atom Fe (gth-dzvp-molopt-sr).

+++ gpu4pyscf/lib/pbc/nr_eval_gto.cu
@@ -301,7 +301,7 @@ static void _cart_deriv0_kernel(double *out, PBCIntEnvVars envs, double *grids,
     }
     int *ao_loc = envs.ao_loc;
     int nf = (li + 1) * (li + 2) / 2;
-    out += ao_loc[bas_id] * ngrids;
+    out += (size_t)ao_loc[bas_id] * ngrids;
     for (int n = 0; n < n_cart_max; ++n) {
         if (n >= nf) break;
         out[n*ngrids+grid_id] = gto[n];
@@ -518,10 +518,10 @@ static void _cart_deriv1_kernel(double *out, PBCIntEnvVars envs, double *grids,
     }
     int *ao_loc = envs.ao_loc;
     int nf = (li + 1) * (li + 2) / 2;
-    out += ao_loc[bas_id] * ngrids + grid_id;
-    double *outx = out + 1 * nao * ngrids;
-    double *outy = out + 2 * nao * ngrids;
-    double *outz = out + 3 * nao * ngrids;
+    out += (size_t)ao_loc[bas_id] * ngrids + grid_id;
+    double *outx = out + 1 * (size_t)nao * ngrids;
+    double *outy = out + 2 * (size_t)nao * ngrids;
+    double *outz = out + 3 * (size_t)nao * ngrids;
     for (int n = 0; n < n_cart_max; ++n) {
         if (n >= nf) break;
         out [n*ngrids] = gto [n];
@@ -623,11 +623,11 @@ static void _cart_ip2_kernel(double *out, PBCIntEnvVars envs, double *grids,
 
     int *ao_loc = envs.ao_loc;
     int nf = (li + 1) * (li + 2) / 2;
-    double *out_x = out + ao_loc[bas_id] * ngrids + grid_id;
+    double *out_x = out + (size_t)ao_loc[bas_id] * ngrids + grid_id;
     for (int n = 0; n < n_cart_max; ++n) {
         if (n >= nf) break;
         for (int ix = 0; ix < 6; ++ix) {
-            out_x[(ix*nao+n)*ngrids] = gto[n*6+ix];
+            out_x[((size_t)ix*nao+n)*ngrids] = gto[n*6+ix];
         }
     }
 }
@@ -724,7 +724,7 @@ static void _sph_deriv0_kernel(double *out, PBCIntEnvVars envs, double *grids,
         }
     }
     int *ao_loc = envs.ao_loc;
-    out += ao_loc[bas_id] * ngrids;
+    out += (size_t)ao_loc[bas_id] * ngrids;
     switch (li) {
     case 0:
         out[grid_id] = gto[0];
@@ -967,11 +967,11 @@ static void _sph_deriv1_kernel(double *out, PBCIntEnvVars envs, double *grids,
         } }
     }
     int *ao_loc = envs.ao_loc;
-    out += ao_loc[bas_id] * ngrids + grid_id;
+    out += (size_t)ao_loc[bas_id] * ngrids + grid_id;
     switch (li) {
     case 0:
         for (int n = 0; n < 4; ++n) {
-            out[n * nao * ngrids] = gto[n];
+            out[n * (size_t)nao * ngrids] = gto[n];
         }
         break;
     case 1:
@@ -1108,11 +1108,11 @@ static void _sph_ip2_kernel(double *out, PBCIntEnvVars envs, double *grids,
     }
 
     int *ao_loc = envs.ao_loc;
-    out += ao_loc[bas_id] * ngrids + grid_id;
+    out += (size_t)ao_loc[bas_id] * ngrids + grid_id;
     switch (li) {
     case 0:
         for (int n = 0; n < 6; ++n) {
-            out[n * nao * ngrids] = gto[n];
+            out[n * (size_t)nao * ngrids] = gto[n];
         }
         break;
     case 1:
@@ -1378,7 +1378,7 @@ static void _cart_deriv0_strain_tensor_kernel(
     if (li < 4) {
         int *ao_loc = envs.ao_loc;
         int nf = (li + 1) * (li + 2) / 2;
-        out += ao_loc[bas_id] * ngrids + grid_id;
+        out += (size_t)ao_loc[bas_id] * ngrids + grid_id;
         size_t naog = nao * ngrids;
         for (int n = 0; n < 10; ++n) {
             if (n >= nf) break;
@@ -1389,7 +1389,7 @@ static void _cart_deriv0_strain_tensor_kernel(
     } else {
         int *ao_loc = envs.ao_loc;
         int nf = (li + 1) * (li + 2) / 2;
-        out += ao_loc[bas_id] * ngrids + grid_id;
+        out += (size_t)ao_loc[bas_id] * ngrids + grid_id;
         size_t naog = nao * ngrids;
         for (int n = 0; n < 15; ++n) {
             if (n >= nf) break;
@@ -1468,7 +1468,7 @@ static void _cart_deriv1_strain_tensor_kernel(
     double _rcut = rcut[bas_id % envs.cell0_nbas];
     double rrcutoff = _rcut * _rcut;
     int *ao_loc = envs.ao_loc;
-    out += ao_loc[bas_id] * ngrids + grid_id;
+    out += (size_t)ao_loc[bas_id] * ngrids + grid_id;
     int nimgs = envs.nimgs;
 
     switch (li) {
